Renamed post type file for consistency
